My lazy scripts
abcbuild—a bash script to build a c file in a directory with
directions. The most important part is every file will add a prefix of
the name of the file, the author and date that created.
abclog— a bash script to create a log file with date.
abcmain— a bash script for me to add main function.
All these scripts were added to $basc director.(bash script)
“abc” prefix of all the name is for easy to locate when I want to build
a new project.